COMP: you guys should try 64bits...
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=7925
git-svn-id: http://svn.slicer.org/Slicer4/trunk@7925 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkCacheManager.cxx
@@ -145,7 +145,7 @@ void vtkCacheManager::SetRemoteCacheFreeBufferSize ( int val )
 void vtkCacheManager::SetRemoteCacheDirectory (const char *dir )
 {
   std::string dirstring = dir ;
-  int len = dirstring.size();
+  int len = (int)dirstring.size();
 
   if ( len > 0 )
     {
