+++ Libs/MRML/vtkMRMLStorableNode.cxx
@@ -133,7 +133,7 @@ void vtkMRMLStorableNode::ReadXMLAttributes(const char** atts)
         //--- now pull apart individual tags
         if ( tags.c_str() != NULL )
           {
-          i = tags.find("=");
+          i = (unsigned int)tags.find("=");
           if ( i != std::string::npos)
             {
             kwd = tags.substr(0, i);
